requirements.txt
fastapi
uvicorn[standard]
python-multipart
ultralytics
torch
pillow